// src/app/ui/portfolio/new-components/DeFiPositionCard/DeFiPositionCard.types.ts
import type { TypographyProps } from '@mui/material/Typography';
import type { TFunction } from 'i18next';
import type { ColumnDefinition } from '@/components/core/ColumnTable/ColumnTable.types';
import type { DefiPosition, DefiToken } from '@/types/jumper-backend';

export interface DeFiPositionCardProps {
  defiPositions?: DefiPosition[];
  onSelect?: (defiPosition: DefiPosition) => void;
  isLoading?: boolean;
}

export interface RenderCellProps<T = DefiToken> {
  item: T;
  titleVariant: TypographyProps['variant'];
  descriptionVariant: TypographyProps['variant'];
  t: TFunction;
  isMobile: boolean;
}

export interface TableSection<T> {
  id: string;
  type: 'supply' | 'borrow' | 'rewards';
  data: T[];
  columns: ColumnDefinition<T>[];
  showHeader: boolean;
}

export type EnhancedDefiTokenWithPositionData = DefiToken &
  Pick<DefiPosition, 'latest' | 'earn'>;

export type SupplySection = TableSection<EnhancedDefiTokenWithPositionData>;
export type BorrowSection = TableSection<EnhancedDefiTokenWithPositionData>;
export type RewardSection = TableSection<EnhancedDefiTokenWithPositionData>;

export type Section = SupplySection | BorrowSection | RewardSection;

export interface PositionGroup {
  position: DefiPosition;
  sections: Section[];
}